requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
fontawesomefree==5.15.4
djangorestframework